Fix passthru and cookies
