+++ spec/cassettes/PagesController/when_visiting_as_a_guest/does_not_have_a_link_to_the_admin_section.yml
@@ -39,12 +39,12 @@ http_interactions:
       Cache-Control:
       - max-age=0, private, must-revalidate
       X-Request-Id:
-      - 1ce20da8-9926-40f3-a082-83511a0bb188
+      - bd57b2e1-2b31-4955-a110-133bdbba7e4f
       X-Runtime:
-      - '0.028889'
+      - '0.019639'
       Server-Timing:
-      - start_processing.action_controller;dur=0.01, sql.active_record;dur=1.19, instantiation.active_record;dur=0.20,
-        process_action.action_controller;dur=6.58
+      - start_processing.action_controller;dur=0.01, sql.active_record;dur=0.98, instantiation.active_record;dur=0.19,
+        process_action.action_controller;dur=5.42
       Content-Length:
       - '409'
     body:
@@ -54,7 +54,7 @@ http_interactions:
         author"},{"id":4,"saved_scenario_id":29,"owner_id":1,"group":"national","title_en":"Super
         scenario for 2030","title_nl":"Super scenario for 2030","version":"latest","end_year":2030,"author":"No
         author"}]}'
-  recorded_at: Tue, 17 Dec 2024 11:43:43 GMT
+  recorded_at: Fri, 10 Jan 2025 10:44:54 GMT
 - request:
     method: get
     uri: http://localhost:3002/api/v1/versions.json
@@ -86,21 +86,21 @@ http_interactions:
       Content-Type:
       - application/json; charset=utf-8
       Etag:
-      - W/"3827ac2c7c96d926b106dcb1102a6dca"
+      - W/"b333f2b38e283af4a4b94a9463c73f73"
       Cache-Control:
       - max-age=0, private, must-revalidate
       X-Request-Id:
-      - b65aa9c4-1f29-4c4f-8c4e-28872a890c46
+      - 36fa8bce-b171-401c-be24-6657cc99ba32
       X-Runtime:
-      - '0.019633'
+      - '0.013402'
       Server-Timing:
-      - start_processing.action_controller;dur=0.01, process_action.action_controller;dur=0.20
+      - start_processing.action_controller;dur=0.01, process_action.action_controller;dur=0.23
       Vary:
       - Origin
       Content-Length:
-      - '212'
+      - '416'
     body:
       encoding: UTF-8
-      string: '{"versions":[{"tag":"latest","url":"https://energytransitionmodel.com"},{"tag":"stable.01","url":"https://stable.energytransitionmodel.com"},{"tag":"stable.02","url":"https://stable2.energytransitionmodel.com"}]}'
-  recorded_at: Tue, 17 Dec 2024 11:43:43 GMT
+      string: '{"versions":[{"tag":"latest","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"},{"tag":"stable.01","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"},{"tag":"stable.02","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"}]}'
+  recorded_at: Fri, 10 Jan 2025 10:44:54 GMT
 recorded_with: VCR 6.0.0
