+++ spec/cassettes/PagesController/with_an_IE11_user_agent_with_the_allow_unsupported_browser_param_set/renders_the_page.yml
@@ -39,12 +39,12 @@ http_interactions:
       Cache-Control:
       - max-age=0, private, must-revalidate
       X-Request-Id:
-      - 9fec44cd-0533-4428-9e49-ae650292b604
+      - 79e79cf2-adb5-4f31-ac1b-b6c5f338dded
       X-Runtime:
-      - '0.019011'
+      - '0.024794'
       Server-Timing:
-      - start_processing.action_controller;dur=0.01, sql.active_record;dur=1.11, instantiation.active_record;dur=0.21,
-        process_action.action_controller;dur=5.90
+      - start_processing.action_controller;dur=0.01, sql.active_record;dur=1.01, instantiation.active_record;dur=0.19,
+        process_action.action_controller;dur=12.29
       Content-Length:
       - '409'
     body:
@@ -54,7 +54,7 @@ http_interactions:
         author"},{"id":4,"saved_scenario_id":29,"owner_id":1,"group":"national","title_en":"Super
         scenario for 2030","title_nl":"Super scenario for 2030","version":"latest","end_year":2030,"author":"No
         author"}]}'
-  recorded_at: Tue, 17 Dec 2024 11:43:42 GMT
+  recorded_at: Fri, 10 Jan 2025 10:44:53 GMT
 - request:
     method: get
     uri: http://localhost:3002/api/v1/versions.json
@@ -86,21 +86,21 @@ http_interactions:
       Content-Type:
       - application/json; charset=utf-8
       Etag:
-      - W/"3827ac2c7c96d926b106dcb1102a6dca"
+      - W/"b333f2b38e283af4a4b94a9463c73f73"
       Cache-Control:
       - max-age=0, private, must-revalidate
       X-Request-Id:
-      - 07d16d7f-9fd6-4277-b9d1-4966968aeff9
+      - a3c18f54-eb30-48d2-b988-230bedd930d8
       X-Runtime:
-      - '0.027959'
+      - '0.013193'
       Server-Timing:
-      - start_processing.action_controller;dur=0.01, process_action.action_controller;dur=0.18
+      - start_processing.action_controller;dur=0.01, process_action.action_controller;dur=0.21
       Vary:
       - Origin
       Content-Length:
-      - '212'
+      - '416'
     body:
       encoding: UTF-8
-      string: '{"versions":[{"tag":"latest","url":"https://energytransitionmodel.com"},{"tag":"stable.01","url":"https://stable.energytransitionmodel.com"},{"tag":"stable.02","url":"https://stable2.energytransitionmodel.com"}]}'
-  recorded_at: Tue, 17 Dec 2024 11:43:42 GMT
+      string: '{"versions":[{"tag":"latest","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"},{"tag":"stable.01","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"},{"tag":"stable.02","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"}]}'
+  recorded_at: Fri, 10 Jan 2025 10:44:54 GMT
 recorded_with: VCR 6.0.0
