+++ spec/cassettes/SavedScenariosController/GET_new/renders_the_new_saved_scenario_form.yml
@@ -31,21 +31,21 @@ http_interactions:
       Content-Type:
       - application/json; charset=utf-8
       Etag:
-      - W/"3827ac2c7c96d926b106dcb1102a6dca"
+      - W/"b333f2b38e283af4a4b94a9463c73f73"
       Cache-Control:
       - max-age=0, private, must-revalidate
       X-Request-Id:
-      - 751e2128-3af8-42d9-a065-7003d1b17b1c
+      - 698004a6-c766-4cb3-9db2-b2f18cc53b9d
       X-Runtime:
-      - '0.016512'
+      - '0.061736'
       Server-Timing:
-      - start_processing.action_controller;dur=0.03, process_action.action_controller;dur=0.33
+      - start_processing.action_controller;dur=0.01, process_action.action_controller;dur=0.22
       Vary:
       - Origin
       Content-Length:
-      - '212'
+      - '416'
     body:
       encoding: UTF-8
-      string: '{"versions":[{"tag":"latest","url":"https://energytransitionmodel.com"},{"tag":"stable.01","url":"https://stable.energytransitionmodel.com"},{"tag":"stable.02","url":"https://stable2.energytransitionmodel.com"}]}'
-  recorded_at: Tue, 17 Dec 2024 11:04:05 GMT
+      string: '{"versions":[{"tag":"latest","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"},{"tag":"stable.01","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"},{"tag":"stable.02","model_url":"http://localhost:3001","engine_url":"http://localhost:3000","collections_url":"http://localhost:3005"}]}'
+  recorded_at: Fri, 10 Jan 2025 10:44:51 GMT
 recorded_with: VCR 6.0.0
